Fixed bugs with Buffs, Sockets, PlayerNames, Mods, Minions, and the GetEntityById function.  We were incorrectly reading INTs instead of longs and getting bad data from time to time.

+++ src/Models/EntityListWrapper.cs
@@ -105,7 +105,7 @@ namespace PoeHUD.Models
             }
         }
 
-        public EntityWrapper GetEntityById(int id)
+        public EntityWrapper GetEntityById(long id)
         {
             EntityWrapper result;
             return entityCache.TryGetValue(id, out result) ? result : null;
